--- web/html/src/manager/login/index.js
@@ -1,3 +1,0 @@
-export default {
-  'login/login': () => import('./login.renderer')
-}
