+++ web/html/src/manager/login/index.ts
@@ -1,0 +1,3 @@
+export default {
+    "login/login": () => import("./login.renderer"),
+};
